backend/requirements.txt
# FastAPI Framework
 fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Database Utilities
sqlalchemy-utils==0.41.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AWS Services
boto3==1.34.0
botocore==1.34.0

# AI Services
anthropic==0.34.0

# NLP & ML
spacy==3.7.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2

# HTTP Clients
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Environment & Configuration
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-httpx==0.26.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3